fix: column names
refactor the project structure
